requirements.txt
# Flask Framework
Flask==3.1.1
Werkzeug==3.1.3

# Database
mysql-connector-python==9.3.0

# Environment Variables
python-dotenv==1.0.0

# Core Dependencies (automatically installed with Flask)
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0

# Production Server (uncomment for production)
# gunicorn==21.2.0